Remove redundant information
Files now only contain the fastd key and no longer the node's name,
which is already visible in the file's name. Also the creation date of
the file is no longer needed in the text file, since this is managed via
git.

The command used to modify the files was:
 for i in *; do tail -1 "$i" | tee "$i"; done